Update harvestPatrolLog variables to match EMR variables
They are all specified in all-caps by convention. Additionally, it
appeared that BoxHarvestBotEmail, BoxHarvestBotUserID, BoxDevpToken were
unused within the codebase, so I have removed them.

Also, it appears that the boxsdk library is able to receive the RSA key
as a string, which would streamline the process (in comparison to having
to write it out to a file). Let's try this and see if it works.